extract resprocessing code out of parser.js
